+++ src/activity/Activity.js
@@ -101,6 +101,7 @@ export function Activity(Behaviour, activityDef, context) {
     isForCompensation,
     attachedTo,
     isTransaction: activityDef.isTransaction,
+    isAdHoc: activityDef.isAdHoc,
     isParallelJoin,
     isParallelGateway: activityDef.isParallelGateway,
     isStartEvent: !!activityDef.isStartEvent,
@@ -195,6 +196,11 @@ Object.defineProperties(Activity.prototype, {
   isTransaction: {
     get() {
       return this[K_FLAGS].isTransaction;
+    },
+  },
+  isAdHoc: {
+    get() {
+      return this[K_FLAGS].isAdHoc;
     },
   },
   isMultiInstance: {